requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
dnspython==2.7.0
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
motor==3.7.1
mpmath==1.3.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-file==2.1.0
safetensors==0.5.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
